# docs/contracts/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetail": {
        "description": "Detalle de un campo inválido (solo en errores de validación).",
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Page_ReviewResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ReviewResponse]",
        "type": "object"
      },
...
      "ReviewCreate": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "id_restaurant": {
            "title": "Id Restaurant",
            "type": "integer"
          },
          "photo_gallery": {
            "description": "URLs de fotos",
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Photo Gallery",
            "type": "array"
          },
          "rating_atmosphere": {
            "description": "De 1 a 5",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating Atmosphere",
            "type": "integer"
          },
          "rating_food": {
            "description": "De 1 a 5",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating Food",
            "type": "integer"
          },
          "rating_service": {
            "description": "De 1 a 5",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating Service",
            "type": "integer"
          }
        },
        "required": [
          "id_restaurant",
          "rating_food",
          "rating_service",
          "rating_atmosphere"
        ],
        "title": "ReviewCreate",
        "type": "object"
      },
...
      "ReviewResponse": {
        "properties": {
          "client_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Del servicio Identity",
            "title": "Client Name"
          },
          "client_photo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Del servicio Identity",
            "title": "Client Photo"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id_client": {
            "title": "Id Client",
            "type": "integer"
          },
          "id_restaurant": {
            "title": "Id Restaurant",
            "type": "integer"
          },
          "id_review": {
            "title": "Id Review",
            "type": "integer"
          },
          "photo_gallery": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Photo Gallery",
            "type": "array"
          },
          "rating_atmosphere": {
            "title": "Rating Atmosphere",
            "type": "integer"
          },
          "rating_food": {
            "title": "Rating Food",
            "type": "integer"
          },
          "rating_service": {
            "title": "Rating Service",
            "type": "integer"
          },
          "restaurant_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Del servicio Catalog",
            "title": "Restaurant Name"
          },
          "status": {
            "enum": [
              "Pending",
              "Approved",
              "Rejected"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id_review",
          "id_client",
          "id_restaurant",
          "rating_food",
          "rating_service",
          "rating_atmosphere",
          "status",
          "created_at"
        ],
        "title": "ReviewResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "dev@crave-app.com",
      "name": "Equipo Crave"
    },
    "description": "Backend de **Crave**, la app para descubrir y reseñar restaurantes. \n\n## Autenticación\nLos endpoints protegidos requieren un header:\n`Authorization: Bearer <tu_token_jwt>`\n\nObten tu token haciendo `POST /api/v1/auth/login`.\n\n## Errores\nTodos los errores responden `{\"error\": {\"code\": \"...\", \"message\": \"...\"}}`.",
    "title": "Crave App API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/reviews": {
      "post": {
        "description": "Publica una reseña y actualiza la calificación del restaurante en Catalog.\n\nEl estado inicial lo fija `REVIEW_DEFAULT_STATUS`: `Approved` (por defecto)\no `Pending` si las reseñas pasan por moderación. Solo las aprobadas se\nmuestran en el restaurante y cuentan para su calificación. Cada cliente\npuede reseñar un restaurante una sola vez.",
        "operationId": "create_review_api_v1_reviews_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "El cuerpo de la petición no se puede leer (p. ej. no es JSON en UTF-8)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Falta el token o es inválido"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Requiere el rol Client, o la reseña es de otro usuario"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Restaurante o reseña no encontrados"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "El cliente ya reseñó este restaurante"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos de entrada inválidos o regla de negocio incumplida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publicar una reseña (Client)",
        "tags": [
          "Reseñas"
        ]
      }
    },
    "/api/v1/reviews/me": {
      "get": {
        "description": "Todas mis reseñas, también las pendientes de moderación, con el nombre del restaurante.",
        "operationId": "get_my_reviews_api_v1_reviews_me_get",
        "parameters": [
          {
            "description": "Elementos por página",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Elementos por página",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Elementos a saltar",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Elementos a saltar",
              "maximum": 2147483647,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ReviewResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "El cuerpo de la petición no se puede leer (p. ej. no es JSON en UTF-8)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Falta el token o es inválido"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos de entrada inválidos o regla de negocio incumplida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mis reseñas",
        "tags": [
          "Reseñas"
        ]
      }
    },
    "/api/v1/reviews/restaurant/{restaurant_id}": {
      "get": {
        "description": "De la más reciente a la más antigua, con el nombre y la foto de cada autor.",
        "operationId": "get_restaurant_reviews_api_v1_reviews_restaurant__restaurant_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "restaurant_id",
            "required": true,
            "schema": {
              "title": "Restaurant Id",
              "type": "integer"
            }
          },
          {
            "description": "Elementos por página",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Elementos por página",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Elementos a saltar",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Elementos a saltar",
              "maximum": 2147483647,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ReviewResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "El cuerpo de la petición no se puede leer (p. ej. no es JSON en UTF-8)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Restaurante o reseña no encontrados"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos de entrada inválidos o regla de negocio incumplida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno del servidor"
          }
        },
        "summary": "Reseñas aprobadas de un restaurante",
        "tags": [
          "Reseñas"
        ]
      }
    },
    "/api/v1/reviews/{review_id}": {
      "delete": {
        "description": "Solo el autor puede eliminarla; la calificación del restaurante se recalcula.",
        "operationId": "delete_review_api_v1_reviews__review_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "El cuerpo de la petición no se puede leer (p. ej. no es JSON en UTF-8)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Falta el token o es inválido"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Requiere el rol Client, o la reseña es de otro usuario"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Restaurante o reseña no encontrados"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datos de entrada inválidos o regla de negocio incumplida"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error interno del servidor"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Eliminar una reseña propia",
        "tags": [
          "Reseñas"
        ]
      }
    },
...
  }
}